requirements.txt
APScheduler~=3.10.4
httpx~=0.27.0
aiofiles~=24.1.0
PyYAML~=6.0.2
rich~=13.7.1
tabulate~=0.9.0
xmltodict~=0.13.0
jieba~=0.42.1
scikit-image~=0.24.0
onnxruntime~=1.18.1
requests~=2.32.3
tortoise-orm~=0.21.5
numpy~=1.26.4
yarl~=1.9.4
pillow~=10.4.0
pydantic~=2.8.2
psutil~=6.0.0
google-generativeai~=0.7.2
edge-tts~=6.1.12
mcstatus~=11.1.1
nonebot2==2.3.2
nonebot-adapter-onebot~=2.4.4
nonebot-plugin-htmlrender~=0.3.3
nonebot2[fastapi]